refactor(chat-core): migrate acceptLocalToolResult into ChatRuntime with typed mesh port (BR14b Lot 10)
